10/10 allocatoins work, a refault fails lookup.

+++ drivers/staging/remote_numa/protocol.h
@@ -33,7 +33,7 @@ enum remote_numa_protocol_version
 };
 
 typedef struct {
-	u8   version : 6;
+	u16  version : 6;
 	u16  type : 10;
 	u16  reserved;
 	u32  main_cookie;
